Fixed useFrameCallback hook not running on UI thread (#3406)
## Description

This is a fix for PR #3372. The `useFrameCallback` hook introduced in that PR was not running on the UI thread, which would make the animations choppy if heavy callculations were running on the JS thread.

## Changes

The `useFrameCallback` hook runs on the UI thread now. This can easily be confirmed by adding `console.log(_WORKLET)` in one of the example callbacks. While running the callback `true` should be written out to the console.

## Checklist

- [ ] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Added TS types tests
- [ ] Added unit / integration tests
- [ ] Updated documentation
- [x] Ensured that CI passes