Avoid multiple rAF calls over JSI when called on the same frame (#3852)
## Summary

This PR overrides requestAnimationFrame on the UI runtime such that it
keeps a list of callbacks for a given frame locally and only calls into
the native requestAnimationFrame once in every frame where the method
gets triggered.

This results in a greatly reduced number of calls that need to happen
over JSI in case where there are multiple components/values being
animated at the same time. This also, apparently, fixes a crash in
BokehExample on android that was due to some race condition issues in
the naitve requestAnimationFrame implementation. Calling it only once in
a given frame resolves that issue.

## Test plan

Run BokehExample on Android with the number of circles bumped up to 300.
The app should not crash and perform well on an emulator (note that 300
will be too high for low-end devices but still the app shouldn't crash).

Co-authored-by: Tomek Zawadzki <tomasz.zawadzki@swmansion.com>