Optimize the way we extract code string and in which we pass it across runtimes (#3930)
## Summary

This PR changes the way we extract code string and additional metadata
for worklets and the way we pass it later on between runtimes such that
we can save on allocating code strings that tend to be relatively big
and waste precious time on allocating those strings.

The scope of the change includes:
1. Changes in plugin code, we no longer assign code string directly to
worklet function objects, instead we extract a metadata object and
define in in the program context (at the main file level). The metadata
includes things that are static, like the code converted to string
(formerly "asString", location, source map and stack details data). We
then reference that metadata object and assign it to the worklet object
under `__initData` field.
2. We update "Core Functions" that need to access code string such that
it fetches the string from a new location – `__initData.code`
3. We refactor the concept of "retaining values" and only conditionally
retain JS values on the UI runtime from the shareable. The purpose of
this change is that we can be sure the initData is persisted on the UI
runtime and not recreated each time the worklet it passed to be executed
on the UI runtime. We add additional flag `shouldRetain` to the
`makeShareableClone` JSI method that allows us to control this behavior.
The refactoring updates the inheritance scheme and makes
`ShareableObjects` etc no longer inherit `RetainingShareable` but
instead we only use `RetainingShareable` when the flag is set. This
change also deletes the previous behavior where we'd use weak objects to
keep the reference to the JS Value as we are unsure about
positive/negative impact of that code (we know it impacts GC times
negatively but not sure if this time is counterbalanced by saving on the
time to convert the values)

## Test plan

Plugin tests needs to be updated. Wanted to put this change out for
others to test.