Make  work with transluscent StatusBar on Android (#3958)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Fixes
https://github.com/software-mansion/react-native-reanimated/issues/3631
Fixes the issue with translucent StatusBar on Android by adding the
option `isStatusBarTranslucentAndroid` in `useAnimatedKeyboard`. It is
the same solution as the one mentioned in the issue:
https://github.com/kirillzyusko/react-native-keyboard-controller/commit/2a7c92f7861e38d2a6d437577a6e3a9a7946a191

## Test plan

There is an excellent example in the issue that reproduces the problem:
https://github.com/ChildishForces/reanimated-bug-repro
I tested on real Android device with Android 12 and RN 71.
I think it would be nice to also test on RN < 70 because the bottom
padding may behave differently.